requirements.txt
amqp==5.3.1
billiard==4.2.4
celery==5.6.2
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
greenlet==3.3.2
h11==0.16.0
Jinja2==3.1.6
kombu==5.6.2
MarkupSafe==3.0.3
packaging==26.0
prompt_toolkit==3.0.52
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
redis==7.2.1
six==1.17.0
SQLAlchemy==2.0.48
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.41.0
vine==5.1.0
wcwidth==0.6.0


# ----------------------------
# Core Web API
# ----------------------------
fastapi
uvicorn[standard]

# ----------------------------
# Async & Background Tasks
# ----------------------------
celery
redis

# ----------------------------
# Database
# ----------------------------
sqlalchemy

# ----------------------------
# Environment Variables
# ----------------------------
python-dotenv

# ----------------------------
# LLM + RAG Stack
# ----------------------------
langchain
langchain-core
langchain-community
langchain-ollama
langchain-google-genai

chromadb
tiktoken

# ----------------------------
# Evaluation
# ----------------------------
ragas
datasets

# ----------------------------
# Embeddings & HTTP
# ----------------------------
requests

# ----------------------------
# Date Handling
# ----------------------------
python-dateutil
dateparser

# ----------------------------
# PDF Processing
# ----------------------------
pypdf

# ----------------------------
# Utilities
# ----------------------------
typing-extensions